selftest/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.4",
    "mariadb>=1.1.14",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
]
